Add support for one-triggers-other order with take-profit leg
This change introduces support for a more advanced order type: a
one-triggers-other order. A request for such an order can ultimately
result in two orders being issued, with the second one (a "leg") only
going into effect once the first one filled. As such, they basically
form an if-then construct. Alpaca supports three types of legs: take-profit,
stop-loss, and stop-limit.
With this change we only add support for take-profit legs. We do so by
introducing a new argument to the order submit command:
--take-profit-price. Using this argument a limit price can be set at
which the leg order will fill (after the initial order filled already).